Python 270 assignment
FILE TRANSFER ASSIGNMENT PART THREE
Users are now asking for a user interface to make using the script easier and more versatile.

Desired features of the UI:

Allow the user to browse and choose a specific folder that will contain the files to be checked daily.

Allow the user to browse and choose a specific folder that will receive the copied files.

Allow the user to manually initiate the 'file check' process that is performed by the script.

Add comments throughout your Python explaining your code.

You have been asked to create this UI.